\name{predict.plrs}
\alias{predict.plrs}

\title{Predict method for \code{plrs} models}

\description{Determine predicted values based on a given plrs model}

\usage{
\method{predict}{plrs}(object, newcghseg, ...)
}

\arguments{
  \item{object}{An object of class \code{\link{plrs-class}}}
  \item{newcghseg}{A vector of new segmented CGH values}
  \item{...}{further arguments}
}

%\details{}

\value{A \code{vector} containing the fitted values.
}

%\references{}

\author{Gwenael G.R. Leday  \email{g.g.r.leday@vu.nl}}

%\note{}
%\seealso{}
%\examples{}
%\keyword{}
